add pod_sync_controller
