+++ cmd/clustertree/cluster-manager/app/manager.go
@@ -45,7 +45,11 @@ func NewClusterManagerCommand(ctx context.Context) (*cobra.Command, error) {
 			if errs := opts.Validate(); len(errs) != 0 {
 				return errs.ToAggregate()
 			}
-			return leaderElectionRun(ctx, opts)
+			// return leaderElectionRun(ctx, opts)
+			if err := run(ctx, opts); err != nil {
+				return err
+			}
+			return nil
 		},
 	}
