+++ pkg/clustertree/cluster-manager/controllers/node_lease_controller.go
@@ -2,8 +2,6 @@ package controllers
 
 import (
 	"context"
-	"fmt"
-	"reflect"
 	"sync"
 	"time"
 
@@ -21,8 +19,6 @@ import (
 
 	kosmosv1alpha1 "github.com/kosmos.io/kosmos/pkg/apis/kosmos/v1alpha1"
 	leafUtils "github.com/kosmos.io/kosmos/pkg/clustertree/cluster-manager/utils"
-	"github.com/kosmos.io/kosmos/pkg/utils/podutils"
-	"github.com/pkg/errors"
 )
 
 const (
@@ -32,7 +28,6 @@ const (
 	DefaultRenewIntervalFraction = 0.25
 
 	DefaultNodeStatusUpdateInterval = 1 * time.Minute
-	DefaultpodStatusUpdateInterval  = 10 * time.Second
 )
 
 type NodeLeaseController struct {
@@ -41,9 +36,8 @@ type NodeLeaseController struct {
 	root             client.Client
 	LeafModelHandler leafUtils.LeafModelHandler
 
-	leaseInterval     time.Duration
-	statusInterval    time.Duration
-	podstatusInterval time.Duration
+	leaseInterval  time.Duration
+	statusInterval time.Duration
 
 	nodes             []*corev1.Node
 	LeafNodeSelectors map[string]kosmosv1alpha1.NodeSelector
@@ -60,7 +54,6 @@ func NewNodeLeaseController(leafClient kubernetes.Interface, root client.Client,
 		LeafNodeSelectors: LeafNodeSelectors,
 		leaseInterval:     getRenewInterval(),
 		statusInterval:    DefaultNodeStatusUpdateInterval,
-		podstatusInterval: DefaultpodStatusUpdateInterval,
 	}
 	return c
 }
@@ -68,7 +61,6 @@ func NewNodeLeaseController(leafClient kubernetes.Interface, root client.Client,
 func (c *NodeLeaseController) Start(ctx context.Context) error {
 	go wait.UntilWithContext(ctx, c.syncLease, c.leaseInterval)
 	go wait.UntilWithContext(ctx, c.syncNodeStatus, c.statusInterval)
-	go wait.UntilWithContext(ctx, c.syncPodStatus, c.podstatusInterval)
 	<-ctx.Done()
 	return nil
 }
@@ -95,74 +87,6 @@ func (c *NodeLeaseController) updateNodeStatus(ctx context.Context, n []*corev1.
 	if err != nil {
 		klog.Errorf("Could not update node status in root cluster,Error: %v", err)
 	}
-	return nil
-}
-
-func (c *NodeLeaseController) syncPodStatus(ctx context.Context) {
-	err := c.updatePodStatus(ctx)
-	if err != nil {
-		klog.Errorf(err.Error())
-	}
-}
-
-func (c *NodeLeaseController) updatePodStatus(ctx context.Context) error {
-	pods, err := c.leafClient.CoreV1().Pods(metav1.NamespaceAll).List(ctx, metav1.ListOptions{})
-	if err != nil {
-		klog.Errorf("Could not list pods in leaf cluster,Error: %v", err)
-		return err
-	}
-
-	var wg sync.WaitGroup
-	errChan := make(chan error, len(pods.Items))
-
-	for _, leafpod := range pods.Items {
-		wg.Add(1)
-		go func(leafpod corev1.Pod) {
-			defer wg.Done()
-			err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-				rootpod := &corev1.Pod{}
-				if err := c.root.Get(ctx, types.NamespacedName{Name: leafpod.Name, Namespace: leafpod.Namespace}, rootpod); err != nil {
-					if apierrors.IsNotFound(err) {
-						klog.Warningf("Pod %s in namespace %s not found in root cluster, skipping...", leafpod.Name, leafpod.Namespace)
-						return nil
-					}
-					return err
-				}
-				if podutils.IsKosmosPod(rootpod) && !reflect.DeepEqual(rootpod.Status, leafpod.Status) {
-					rPodCopy := rootpod.DeepCopy()
-					rPodCopy.Status = leafpod.Status
-					podutils.FitObjectMeta(&rPodCopy.ObjectMeta)
-					if err := c.root.Status().Update(ctx, rPodCopy); err != nil && !apierrors.IsNotFound(err) {
-						klog.V(4).Info(errors.Wrap(err, "error while updating pod status in kubernetes"))
-						return err
-					}
-				}
-				return nil
-			})
-			if err != nil {
-				//klog.Errorf("failed to update pod %s/%s, error: %v", leafpod.Namespace, leafpod.Name, err)
-				errChan <- fmt.Errorf("failed to update pod %s/%s, error: %v", leafpod.Namespace, leafpod.Name, err)
-			}
-		}(leafpod)
-	}
-
-	wg.Wait()
-	close(errChan)
-
-	var taskErr error
-	for err := range errChan {
-
-		if taskErr == nil {
-			taskErr = err
-		} else {
-			taskErr = errors.Wrap(err, taskErr.Error())
-		}
-	}
-
-	if taskErr != nil {
-		return taskErr
-	}
-
 	return nil
 }
